requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.2
PyMuPDF==1.23.16
fpdf==1.7.2
Pillow==10.2.0
python-docx==1.1.0
python-dotenv==1.1.0
docx2pdf==0.1.8
pathlib
mysql-connector-python==8.3.0
openpyxl==3.1.2
python-pptx==0.6.21
pdf2image==1.16.3
reportlab==4.1.0
cryptography==39.0.2